# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A meow API server built with Go, inspired by meow",
        "title": "meow meow API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "meow API Support",
            "url": "https://github.com/your-username/meow",
            "email": "support@meow.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/session/{sessionId}/contacts/avatar": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get contact's profile picture/avatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "summary": "Get contact avatar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Get avatar request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetAvatarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    }
                }
            }
        },
        "/session/{sessionId}/contacts/check": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Check if phone numbers are registered on meow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "summary": "Check contacts on meow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Check contact request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CheckContactRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    }
                }
            }
        },
        "/session/{sessionId}/contacts/info": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get detailed information about contacts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "summary": "Get contact information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Get contact info request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetContactInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    }
                }
            }
        },
        "/session/{sessionId}/contacts/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all contacts from contact's meow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "summary": "Get contacts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    }
                }
            }
        },
        "/session/{sessionId}/contacts/presence": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Set global contact presence (available/unavailable)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "summary": "Set contact presence",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Set presence request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SetContactPresenceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ContactResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AvatarInfo": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "example": "https://pps.meow.net/..."
                },
                "jid": {
                    "type": "string",
                    "example": "5511999999999@s.meow.net"
                },
                "phone": {
                    "type": "string",
                    "example": "5511999999999"
                },
                "picture_id": {
                    "type": "string",
                    "example": "pic_123"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
...
        "dto.CheckContactRequest": {
            "type": "object",
            "required": [
                "phones"
            ],
            "properties": {
                "phones": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"5511999999999\"",
                        " \"5511888888888\"]"
                    ]
                }
            }
        },
...
        "dto.ContactCheckResult": {
            "type": "object",
            "properties": {
                "is_in_meow": {
                    "type": "boolean",
                    "example": true
                },
                "jid": {
                    "type": "string",
                    "example": "5511999999999@s.meow.net"
                },
                "query": {
                    "type": "string",
                    "example": "5511999999999"
                },
                "verified_name": {
                    "type": "string",
                    "example": "João Silva"
                }
            }
        },
        "dto.ContactData": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "check_contacts"
                },
                "avatar": {
                    "$ref": "#/definitions/dto.AvatarInfo"
                },
                "check_results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ContactCheckResult"
                    }
                },
                "contact_infos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ContactInfo"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
        "dto.ContactErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_PHONE"
                },
                "details": {
                    "type": "string",
                    "example": "Phone number must include country code"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid phone number format"
                }
            }
        },
        "dto.ContactInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://..."
                },
                "business_name": {
                    "type": "string",
                    "example": "Empresa João"
                },
                "device_count": {
                    "type": "integer",
                    "example": 2
                },
                "display_name": {
                    "type": "string",
                    "example": "João Silva"
                },
                "is_blocked": {
                    "type": "boolean",
                    "example": false
                },
                "is_muted": {
                    "type": "boolean",
                    "example": false
                },
                "jid": {
                    "type": "string",
                    "example": "5511999999999@s.meow.net"
                },
                "name": {
                    "type": "string",
                    "example": "João Silva"
                },
                "notify": {
                    "type": "string",
                    "example": "João"
                },
                "picture_id": {
                    "type": "string",
                    "example": "pic_123"
                },
                "push_name": {
                    "type": "string",
                    "example": "João"
                },
                "status": {
                    "type": "string",
                    "example": "Disponível"
                },
                "verified_name": {
                    "type": "string",
                    "example": "João Silva Empresa"
                }
            }
        },
...
        "dto.ContactResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/dto.ContactData"
                },
                "error": {
                    "$ref": "#/definitions/dto.ContactErrorResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.ContactsData": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "get_contacts"
                },
                "contacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ContactInfo"
                    }
                },
                "count": {
                    "type": "integer",
                    "example": 10
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
        "dto.ContactsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/dto.ContactsData"
                },
                "error": {
                    "$ref": "#/definitions/dto.ContactErrorResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.GetAvatarRequest": {
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "5511999999999"
                }
            }
        },
...
        "dto.GetContactInfoRequest": {
            "type": "object",
            "required": [
                "phones"
            ],
            "properties": {
                "phones": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"5511999999999\"",
                        " \"5511888888888\"]"
                    ]
                }
            }
        },
...
        "dto.SetContactPresenceRequest": {
            "type": "object",
            "required": [
                "state"
            ],
            "properties": {
                "state": {
                    "type": "string",
                    "example": "available"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication. Simply provide your API key directly: \"YOUR_API_KEY\". The system automatically detects if it's a Global API Key (can access all sessions and session management) or a Session-specific API Key (can only access the specific session it belongs to).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}